+++ src/models/userModel.py
@@ -15,12 +15,21 @@ class userModel(object):
 		self._user_token = ''
 		self._enabled = 0
 
-
 	def init(self, data):
-		self.username = data['username'] if data.has_key('username') else ''
+		self.username = data['user_name'] if data.has_key('user_name') else ''
 		self.email = data['email'] if data.has_key('email') else ''
 		self.password = data['password']
 		self.user_token = data['user_token'] if data.has_key('user_token') else ''
+
+	# convert object to printable dictionary
+	def serialize(self):
+		data = {}
+		data['user_name'] = self.username
+		data['email'] = self.email
+		data['password'] = self.password
+		data['user_token'] = self.user_token
+		
+		return data
 
 	@property
 	def username(self):
